replace lakitu cam with custom camera mode
Replaced lakitu cam with a custom camera mode. Inspired by Aglab's parallel lakitu cam, this mode is a modified version of the return_boss_fight_camera_yaw camera. It sports free horizontal rotation (rather than incremental with each C button press), ignores wall collisions, and never rotates when not instructed by the player to do so. Additionally, this camera shifts its position down a bit when mario is hanging from a ceiling, in order to assist with visibility (this part is subject to change).